refactor(EST-2124-BE): clean up communication schemas and reorganize search-related models
