+++ src/knowledge_new/enums.py
@@ -1,4 +1,5 @@
 from enum import StrEnum
+from src.shared.enums import GraphSearchMethodEnum, RAGStrategy
 
 __all__ = [
     "ChunkStrategyEnum",
@@ -10,11 +11,6 @@ __all__ = [
     "IndexStatusEnum",
     "RAGStrategy",
 ]
-
-
-class RAGStrategy(StrEnum):
-    NAIVE = "naive"
-    GRAPH = "graph"
 
 
 class ChunkStrategyEnum(StrEnum):
@@ -52,13 +48,6 @@ class EmbedderProviderEnum(StrEnum):
     MISTRAL = "mistral"
     OPENAI = "openai"
     TOGETHER_AI = "together_ai"
-
-
-class GraphSearchMethodEnum(StrEnum):
-    BASIC = "basic"
-    LOCAL = "local"
-    GLOBAL = "global"
-    DRIFT = "drift"
 
 
 class IndexStatusEnum(StrEnum):
